safari-534.53-branch/LayoutTests/fast/dom/DeviceMotion/window-property-expected.txt

Tests that the window.DeviceMotionEvent and window.ondevicemotion properties are present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof window.DeviceMotionEvent == 'object' is true
PASS typeof window.DeviceMotionEvent == 'function' is false
PASS hasDeviceMotionEventProperty() is true
PASS 'DeviceMotionEvent' in window is true
PASS window.hasOwnProperty('DeviceMotionEvent') is true
PASS typeof window.ondevicemotion == 'object' is true
PASS hasOnDeviceMotionProperty() is true
PASS 'ondevicemotion' in window is true
PASS window.hasOwnProperty('ondevicemotion') is true
PASS successfullyParsed is true

TEST COMPLETE
